Fix conda build.
