+++ skillmodels/config.py
@@ -1,4 +1,3 @@
 from pathlib import Path
 
-
 TEST_DIR = Path(__file__).resolve().parent / "tests"
