--- skillmodels/tests/conftest.py
@@ -1,19 +1,0 @@
-import pandas as pd
-import pytest
-import yaml
-
-from skillmodels.config import TEST_DIR
-
-
-@pytest.fixture
-def model2():
-    with open(TEST_DIR / "model2.yaml") as y:
-        model_dict = yaml.load(y, Loader=yaml.FullLoader)
-    return model_dict
-
-
-@pytest.fixture
-def model2_data():
-    data = pd.read_stata(TEST_DIR / "model2_simulated_data.dta")
-    data = data.set_index(["caseid", "period"])
-    return data
