+++ skillmodels/tests/test_simulate_data.py
@@ -4,6 +4,7 @@ from pathlib import Path
 import numpy as np
 import pandas as pd
 import pytest
+import yaml
 from numpy.testing import assert_array_almost_equal as aaae
 from pandas.testing import assert_frame_equal
 
@@ -12,7 +13,17 @@ from skillmodels.simulate_data import measurements_from_states
 from skillmodels.simulate_data import next_period_states
 from skillmodels.simulate_data import simulate_dataset
 
+
+# importing the TEST_DIR from config does not work for test run in conda build
 TEST_DIR = Path(__file__).parent.resolve()
+
+
+@pytest.fixture
+def model2():
+    with open(TEST_DIR / "model2.yaml") as y:
+        model_dict = yaml.load(y, Loader=yaml.FullLoader)
+    return model_dict
+
 
 # =======================================================
 # test that simulate_dataset works with the example model
